requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
_anaconda_depends=2019.03=py37_0
_pytorch_select=0.1=cpu_0
_tflow_select=2.3.0=mkl
absl-py=0.9.0=py37_0
alabaster=0.7.12=py37_0
anaconda=custom=py37_1
anaconda-client=1.7.2=py37_0
anaconda-project=0.8.4=py_0
appnope=0.1.0=py37_0
appscript=1.1.0=py37h1de35cc_0
asn1crypto=1.3.0=py37_0
astor=0.8.0=py37_0
astroid=2.3.3=py37_0
astropy=4.0=py37h1de35cc_0
atomicwrites=1.3.0=py37_1
attrs=19.3.0=py_0
babel=2.8.0=py_0
backcall=0.1.0=py37_0
backports=1.0=py_2
backports.os=0.1.1=py37_0
backports.shutil_get_terminal_size=1.0.0=py37_2
beautifulsoup4=4.8.2=py37_0
bitarray=1.2.1=py37h1de35cc_0
bkcharts=0.2=py37_0
blas=1.0=mkl
bleach=3.1.0=py37_0
blosc=1.16.3=hd9629dc_0
bokeh=1.4.0=py37_0
boto=2.49.0=py37_0
bottleneck=1.3.1=py37h1d22016_0
bzip2=1.0.8=h1de35cc_0
c-ares=1.15.0=h1de35cc_1001
ca-certificates=2020.1.1=0
certifi=2019.11.28=py37_0
cffi=1.14.0=py37hb5b8e2f_0
chardet=3.0.4=py37_1003
click=7.0=py37_0
cloudpickle=1.3.0=py_0
clyent=1.2.2=py37_1
colorama=0.4.3=py_0
contextlib2=0.6.0.post1=py_0
cryptography=2.8=py37ha12b0ac_0
curl=7.68.0=ha441bb4_0
cycler=0.10.0=py37_0
cython=0.29.15=py37h0a44026_0
cytoolz=0.10.1=py37h1de35cc_0
dask=2.10.1=py_0
dask-core=2.10.1=py_0
dbus=1.13.12=h90a0687_0
decorator=4.4.1=py_0
defusedxml=0.6.0=py_0
distributed=2.10.0=py_0
docutils=0.16=py37_0
entrypoints=0.3=py37_0
et_xmlfile=1.0.1=py37_0
expat=2.2.6=h0a44026_0
fastcache=1.1.0=py37h1de35cc_0
filelock=3.0.12=py_0
fire=0.2.1=pypi_0
flask=1.1.1=py_0
freetype=2.9.1=hb4e5f40_0
fsspec=0.6.2=py_0
gast=0.3.3=py_0
get_terminal_size=1.0.0=h7520d66_0
gettext=0.19.8.1=h15daf44_3
gevent=1.4.0=py37h1de35cc_0
glib=2.63.1=hd977a24_0
glob2=0.7=py_0
gmp=6.1.2=hb37e062_1
gmpy2=2.0.8=py37h6ef4df4_2
greenlet=0.4.15=py37h1de35cc_0
grpcio=1.16.1=py37h044775b_1
h5py=2.10.0=py37h3134771_0
hdf5=1.10.4=hfa1e0ec_0
heapdict=1.0.1=py_0
html5lib=1.0.1=py37_0
hypothesis=5.4.1=py_0
icu=58.2=h4b95b61_1
idna=2.8=py37_0
imageio=2.6.1=py37_0
imagesize=1.2.0=py_0
importlib_metadata=1.5.0=py37_0
intel-openmp=2019.4=233
ipykernel=5.1.4=py37h39e3cac_0
ipython=7.12.0=py37h5ca1d4c_0
ipython_genutils=0.2.0=py37_0
ipywidgets=7.5.1=py_0
isort=4.3.21=py37_0
itsdangerous=1.1.0=py37_0
jbig=2.1=h4d881f8_0
jdcal=1.4.1=py_0
jedi=0.16.0=py37_0
jinja2=2.11.1=py_0
joblib=0.14.1=py_0
jpeg=9b=he5867d9_2
json5=0.9.1=py_0
jsonpatch=1.25=pypi_0
jsonpointer=2.0=pypi_0
jsonschema=3.2.0=py37_0
jupyter=1.0.0=py37_7
jupyter_client=5.3.4=py37_0
jupyter_console=6.1.0=py_0
jupyter_core=4.6.1=py37_0
jupyterlab=1.2.6=pyhf63ae98_0
jupyterlab_server=1.0.6=py_0
keras=2.2.4=0
keras-applications=1.0.8=py_0
keras-base=2.2.4=py37_0
keras-preprocessing=1.1.0=py_1
keyring=21.1.0=py37_0
kiwisolver=1.1.0=py37h0a44026_0
krb5=1.17.1=hddcf347_0
lazy-object-proxy=1.4.3=py37h1de35cc_0
libarchive=3.3.3=h786848e_5
libcurl=7.68.0=h051b688_0
libcxx=4.0.1=hcfea43d_1
libcxxabi=4.0.1=hcfea43d_1
libedit=3.1.20181209=hb402a30_0
libffi=3.2.1=h475c297_4
libgfortran=3.0.1=h93005f0_2
libiconv=1.15=hdd342a3_7
liblief=0.9.0=h2a1bed3_2
libpng=1.6.37=ha441bb4_0
libprotobuf=3.11.3=hd9629dc_0
libsodium=1.0.16=h3efe00b_0
libssh2=1.8.2=ha12b0ac_0
libtiff=4.1.0=hcb84e12_0
libxml2=2.9.9=hf6e021a_1
libxslt=1.1.33=h33a18ac_0
llvm-openmp=4.0.1=hcfea43d_1
llvmlite=0.31.0=py37h1341992_0
locket=0.2.0=py37_1
lxml=4.5.0=py37hef8c89e_0
lz4-c=1.8.1.2=h1de35cc_0
lzo=2.10=h362108e_2
markdown=3.1.1=py37_0
markupsafe=1.1.1=py37h1de35cc_0
matplotlib=3.1.3=py37_0
matplotlib-base=3.1.3=py37h9aa3819_0
mccabe=0.6.1=py37_1
mistune=0.8.4=py37h1de35cc_0
mkl=2019.4=233
mkl-service=2.3.0=py37hfbe908c_0
mkl_fft=1.0.15=py37h5e564d8_0
mkl_random=1.1.0=py37ha771720_0
mock=4.0.1=py_0
more-itertools=8.2.0=py_0
mpc=1.1.0=h6ef4df4_1
mpfr=4.0.1=h3018a27_3
mpmath=1.1.0=py37_0
msgpack-python=0.6.1=py37h04f5b5a_1
multipledispatch=0.6.0=py37_0
nbconvert=5.6.1=py37_0
nbformat=5.0.4=py_0
ncurses=6.1=h0a44026_1
networkx=2.4=py_0
ninja=1.9.0=py37h04f5b5a_0
nltk=3.4.5=py37_0
nose=1.3.7=py37_2
notebook=6.0.3=py37_0
numba=0.48.0=py37h6c726b0_0
numexpr=2.7.1=py37hce01a72_0
numpy=1.18.1=py37h7241aed_0
numpy-base=1.18.1=py37h6575580_1
numpydoc=0.9.2=py_0
olefile=0.46=py37_0
openpyxl=3.0.3=py_0
openssl=1.1.1d=h1de35cc_4
packaging=20.1=py_0
pandas=1.0.1=py37h6c726b0_0
pandoc=2.2.3.2=0
pandocfilters=1.4.2=py37_1
parso=0.6.1=py_0
partd=1.1.0=py_0
path=13.1.0=py37_0
path.py=12.4.0=0
pathlib2=2.3.5=py37_0
patsy=0.5.1=py37_0
pcre=8.43=h0a44026_0
pep8=1.7.1=py37_0
pexpect=4.8.0=py37_0
pickleshare=0.7.5=py37_0
pillow=7.0.0=py37h4655f20_0
pip=20.0.2=py37_1
pkginfo=1.5.0.1=py37_0
pluggy=0.13.1=py37_0
ply=3.11=py37_0
prometheus_client=0.7.1=py_0
prompt_toolkit=3.0.3=py_0
protobuf=3.11.3=py37h0a44026_0
psutil=5.6.7=py37h1de35cc_0
ptyprocess=0.6.0=py37_0
py=1.8.1=py_0
py-lief=0.9.0=py37h1413db1_2
pycodestyle=2.5.0=py37_0
pycosat=0.6.3=py37h1de35cc_0
pycparser=2.19=py37_0
pycrypto=2.6.1=py37h1de35cc_9
pycurl=7.43.0.5=py37ha12b0ac_0
pyflakes=2.1.1=py37_0
pygments=2.5.2=py_0
pylint=2.4.4=py37_0
pyodbc=4.0.30=py37h0a44026_0
pyopenssl=19.1.0=py37_0
pyparsing=2.4.6=py_0
pyqt=5.9.2=py37h655552a_2
pyrsistent=0.15.7=py37h1de35cc_0
pysocks=1.7.1=py37_0
pytables=3.6.1=py37h5bccee9_0
pytest=5.3.5=py37_0
pytest-arraydiff=0.3=py37h39e3cac_0
pytest-astropy=0.8.0=py_0
pytest-astropy-header=0.1.2=py_0
pytest-doctestplus=0.5.0=py_0
pytest-openfiles=0.4.0=py_0
pytest-remotedata=0.3.2=py37_0
python=3.7.6=h359304d_2
python-dateutil=2.8.1=py_0
python-libarchive-c=2.8=py37_13
python.app=2=py37_9
pytorch=1.3.1=cpu_py37h0c87eb2_0
pytz=2019.3=py_0
pywavelets=1.1.1=py37h1de35cc_0
pyyaml=5.3=py37h1de35cc_0
pyzmq=18.1.1=py37h0a44026_0
qt=5.9.7=h468cd18_1
qtawesome=0.6.1=py_0
qtconsole=4.6.0=py_1
qtpy=1.9.0=py_0
readline=7.0=h1de35cc_5
requests=2.22.0=py37_1
rope=0.16.0=py_0
ruamel_yaml=0.15.87=py37h1de35cc_0
scikit-image=0.16.2=py37h6c726b0_0
scikit-learn=0.22.1=py37h27c97d8_0
scipy=1.4.1=py37h9fa6033_0
seaborn=0.10.0=py_0
send2trash=1.5.0=py37_0
setuptools=45.2.0=py37_0
simplegeneric=0.8.1=py37_2
singledispatch=3.4.0.3=py37_0
sip=4.19.8=py37h0a44026_0
six=1.14.0=py37_0
skorch=0.7.0=pypi_0
snappy=1.1.7=he62c110_3
snowballstemmer=2.0.0=py_0
sortedcollections=1.1.2=py37_0
sortedcontainers=2.1.0=py37_0
soupsieve=1.9.5=py37_0
sphinx=2.4.0=py_0
sphinxcontrib=1.0=py37_1
sphinxcontrib-applehelp=1.0.1=py_0
sphinxcontrib-devhelp=1.0.1=py_0
sphinxcontrib-htmlhelp=1.0.2=py_0
sphinxcontrib-jsmath=1.0.1=py_0
sphinxcontrib-qthelp=1.0.2=py_0
sphinxcontrib-serializinghtml=1.1.3=py_0
sphinxcontrib-websupport=1.2.0=py_0
spyder=3.3.6=py37_0
spyder-kernels=0.5.2=py37_0
sqlalchemy=1.3.13=py37h1de35cc_0
sqlite=3.31.1=ha441bb4_0
statsmodels=0.11.0=py37h1de35cc_0
sympy=1.5.1=py37_0
tabulate=0.8.6=pypi_0
tbb=2020.0=h04f5b5a_0
tblib=1.6.0=py_0
tensorboard=1.14.0=py37h80053f4_0
tensorflow=1.14.0=mkl_py37h085be34_0
tensorflow-base=1.14.0=mkl_py37h5a24fda_0
tensorflow-estimator=1.14.0=py_0
termcolor=1.1.0=py37_1
terminado=0.8.3=py37_0
testpath=0.4.4=py_0
tk=8.6.8=ha441bb4_0
toolz=0.10.0=py_0
torchfile=0.1.0=pypi_0
torchnet=0.0.4=pypi_0
tornado=6.0.3=py37h1de35cc_3
tqdm=4.42.1=py_0
traitlets=4.3.3=py37_0
unicodecsv=0.14.1=py37_0
unixodbc=2.3.7=h1de35cc_0
urllib3=1.25.8=py37_0
visdom=0.1.8.9=pypi_0
wcwidth=0.1.8=py_0
webencodings=0.5.1=py37_1
websocket-client=0.57.0=pypi_0
werkzeug=1.0.0=py_0
wheel=0.34.2=py37_0
widgetsnbextension=3.5.1=py37_0
wrapt=1.11.2=py37h1de35cc_0
wurlitzer=2.0.0=py37_0
xlrd=1.2.0=py37_0
xlsxwriter=1.2.7=py_0
xlwings=0.17.1=py37_0
xlwt=1.3.0=py37_0
xz=5.2.4=h1de35cc_4
yaml=0.1.7=hc338f04_2
zeromq=4.3.1=h0a44026_3
zict=1.0.0=py_0
zipp=2.2.0=py_0
zlib=1.2.11=h1de35cc_3
zstd=1.3.7=h5bba6e5_0
